requirements.txt
requests
parsel
pycryptodome
psutil